+++ sdk/storage/storage-blob/recordings/node/containerurl/recording_setaccesspolicy_should_work_when_expiry_and_start_undefined.js
@@ -1,0 +1,99 @@
+let nock = require('nock');
+
+module.exports.testInfo = {"container":"container156635862622505975"}
+
+nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
+  .put('/container156635862622505975')
+  .query(true)
+  .reply(201, "", [
+  'Content-Length',
+  '0',
+  'Last-Modified',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'ETag',
+  '"0x8D725E84882598A"',
+  'Server',
+  'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '5c15bd59-a01e-0047-11d1-57e1fd000000',
+  'x-ms-version',
+  '2018-11-09',
+  'Date',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'Connection',
+  'close'
+]);
+
+
+nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
+  .put('/container156635862622505975', "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?><SignedIdentifiers><SignedIdentifier><Id>MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=</Id><AccessPolicy><Start/><Expiry/><Permission>rwd</Permission></AccessPolicy></SignedIdentifier></SignedIdentifiers>")
+  .query(true)
+  .reply(200, "", [
+  'Content-Length',
+  '0',
+  'Last-Modified',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'ETag',
+  '"0x8D725E848AF6F87"',
+  'Server',
+  'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '73cdf2e7-701e-00c6-03d1-5743aa000000',
+  'x-ms-version',
+  '2018-11-09',
+  'Date',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'Connection',
+  'close'
+]);
+
+
+nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
+  .get('/container156635862622505975')
+  .query(true)
+  .reply(200, "﻿<?xml version=\"1.0\" encoding=\"utf-8\"?><SignedIdentifiers><SignedIdentifier><Id>MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=</Id><AccessPolicy><Permission>rwd</Permission></AccessPolicy></SignedIdentifier></SignedIdentifiers>", [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/xml',
+  'Last-Modified',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'ETag',
+  '"0x8D725E848AF6F87"',
+  'Server',
+  'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '1b4d58c3-c01e-00d4-32d1-5777b6000000',
+  'x-ms-version',
+  '2018-11-09',
+  'x-ms-blob-public-access',
+  'blob',
+  'Access-Control-Expose-Headers',
+  'x-ms-request-id,Server,x-ms-version,x-ms-blob-public-access,Last-Modified,ETag,Content-Length,Date,Transfer-Encoding',
+  'Access-Control-Allow-Origin',
+  '*',
+  'Date',
+  'Wed, 21 Aug 2019 03:33:08 GMT',
+  'Connection',
+  'close'
+]);
+
+
+nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
+  .delete('/container156635862622505975')
+  .query(true)
+  .reply(202, "", [
+  'Content-Length',
+  '0',
+  'Server',
+  'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '7bb45b1a-b01e-0131-03d1-572314000000',
+  'x-ms-version',
+  '2018-11-09',
+  'Date',
+  'Wed, 21 Aug 2019 03:33:07 GMT',
+  'Connection',
+  'close'
+]);
+